source/requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
Django==2.1.7
django-redis==4.10.0
djangorestframework==3.9.2
idna==2.8
kombu==4.4.0
openpyxl==2.6.1
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
